+++ Sources/SwiftMarkdownViewLaTeX/LaTeXMathRenderer.swift
@@ -14,7 +14,7 @@ import SwiftLaTeXView
 ///
 /// $$x = \\frac{-b \\pm \\sqrt{b^2 - 4ac}}{2a}$$
 /// """)
-/// .mathRenderer(LaTeXMathRenderer())
+/// .markdownMathRenderer(LaTeXMathRenderer())
 /// ```
 public struct LaTeXMathRenderer: MathRenderer {